# geoindex: make geometry index bounds a square